\name{STDE}
\alias{STDE}

\title{SE for lrqmm}
\description{
This function writed in "summary.rq" in "quantreg" package but in below used and changed for lrqmm function.
}
\details{This function runs in "lrqmm_m" function.}

\references{
[1]Alavian, S. R. (2019). Creating LRQMM package for predicting the breeding value of animals
by corrected mixed quantile regression (Unpublished master's thesis). Ferdowsi University Of Mashhad. Mashhad. Iran.[Persian].
}
\author{Sayyed Reza Alavian}
