\name{Cow}
\alias{Cow}
\docType{data}
\title{breastfeeding period data of Iranian Holstein cows}
\description{ Cow data include 30 cows with id records, father's record, mother' record, number of Herd and breastfeeding period length.}
\usage{data("Cow")}
\format{
  A data frame with 30 observations on the following 5 variables.
  \describe{
    \item{\code{id}}{ The number form animal record as vector or column matrix}
    \item{\code{sire}}{The number form father's animal record as vector or column matrix}
    \item{\code{dam}}{The number form mother's animal record as vector or column matrix}
    \item{\code{HERD}}{a number of record's Herd}
    \item{\code{LACTLENGHT}}{a numeric vector of breastfeeding period length }
  }
}

\examples{
data(Cow)

}
\keyword{datasets}
